chore: rename file to page
