+++ internal/repository/postgres/page.go
@@ -1,177 +1,187 @@
 package postgres
 
-// import (
-// 	"context"
+import (
+	"context"
 
-// 	"time"
+	"github.com/Stuhub-io/config"
+	"github.com/Stuhub-io/core/domain"
+	store "github.com/Stuhub-io/internal/repository"
+	"github.com/Stuhub-io/internal/repository/model"
+	"github.com/Stuhub-io/utils/docutils"
+	"gorm.io/gorm/clause"
+)
 
-// 	"github.com/Stuhub-io/config"
-// 	"github.com/Stuhub-io/core/domain"
-// 	"github.com/Stuhub-io/core/ports"
-// 	store "github.com/Stuhub-io/internal/repository"
-// 	"github.com/Stuhub-io/internal/repository/model"
-// 	"github.com/Stuhub-io/utils/pageutils"
-// 	"github.com/google/uuid"
-// 	"gorm.io/gorm/clause"
-// )
+type DocRepository struct {
+	store *store.DBStore
+	cfg   config.Config
+}
 
-// type PageRepository struct {
-// 	store *store.DBStore
-// 	cfg   config.Config
-// }
+type NewDocRepositoryParams struct {
+	Cfg   config.Config
+	Store *store.DBStore
+}
 
-// type NewPageRepositoryParams struct {
-// 	Cfg   config.Config
-// 	Store *store.DBStore
-// }
+func NewDocRepository(params NewDocRepositoryParams) *DocRepository {
+	return &DocRepository{
+		store: params.Store,
+		cfg:   params.Cfg,
+	}
+}
 
-// func NewPageRepository(params NewPageRepositoryParams) ports.PageRepository {
-// 	return &PageRepository{
-// 		store: params.Store,
-// 		cfg:   params.Cfg,
-// 	}
-// }
+func (r *DocRepository) List(ctx context.Context, q domain.PageListQuery) ([]domain.Page, *domain.Error) {
+	var pages []model.Page
+	query := r.store.DB().Where("org_pkid = ?", q.OrgPkID)
 
-// func (r *PageRepository) GetPagesBySpacePkID(ctx context.Context, spacePkID int64, excludeArchived bool) ([]domain.Page, *domain.Error) {
-// 	var pages []model.Page
-// 	query := r.store.DB().Where("space_pkid = ?", spacePkID)
-// 	if excludeArchived {
-// 		query = query.Where("archived_at IS NULL")
-// 	}
-// 	err := query.Order("created_at desc").Find(&pages).Error
-// 	if err != nil {
-// 		return nil, domain.ErrDatabaseQuery
-// 	}
+	// Filter by Archived
+	if q.IsArchived != nil {
+		query = query.Where("archived_at IS NULL = ?", !*q.IsArchived)
+	}
 
-// 	var domainPages []domain.Page
-// 	for _, page := range pages {
-// 		domainPages = append(domainPages, *pageutils.TransformPageModelToDomain(page, nil))
-// 	}
+	// Filter By Parent Page
+	if q.ParentPagePkID != nil {
+		query = query.Where("parent_page_pkid = ?", *q.ParentPagePkID)
+	}
 
-// 	return domainPages, nil
-// }
+	if len(q.ViewTypes) > 0 {
+		query = query.Where("view_type IN ?", q.ViewTypes)
+	}
 
-// func (r *PageRepository) DeletePageByPkID(ctx context.Context, pagePkID int64, userPkID int64) (*domain.Page, *domain.Error) {
-// 	var page model.Page
-// 	isDeleted := r.store.DB().Where("pkid = ?", pagePkID).Delete(&page).Error
-// 	if isDeleted != nil {
-// 		return nil, domain.ErrDatabaseDelete
-// 	}
+	query = query.Order("created_at desc").Offset(q.Offset).Limit(q.Limit)
 
-// 	return pageutils.TransformPageModelToDomain(page, nil), nil
-// }
+	err := query.Find(&pages).Error
+	if err != nil {
+		return nil, domain.ErrDatabaseQuery
+	}
 
-// func (r *PageRepository) GetPageByID(ctx context.Context, pageID string) (*domain.Page, *domain.Error) {
-// 	var page model.Page
-// 	err := r.store.DB().Where("id = ?", pageID).First(&page).Error
-// 	if err != nil {
-// 		return nil, domain.ErrDatabaseQuery
-// 	}
-// 	var childPages []model.Page
-// 	rerr := r.store.DB().Where("parent_page_pkid = ?", page.Pkid).Find(&childPages).Error
-// 	if rerr != nil {
-// 		return nil, domain.ErrDatabaseQuery
-// 	}
-// 	childPagesDomain := make([]domain.Page, 0)
-// 	for _, childPage := range childPages {
-// 		childPagesDomain = append(childPagesDomain, *pageutils.TransformPageModelToDomain(childPage, nil))
-// 	}
+	var domainPages []domain.Page = make([]domain.Page, 0, len(pages))
+	for _, page := range pages {
+		domainPages = append(domainPages, *docutils.TransformPageModelToDomain(page, nil, nil))
+	}
+	return domainPages, nil
+}
 
-// 	return pageutils.TransformPageModelToDomain(page, childPagesDomain), nil
-// }
+func (r *DocRepository) Update(ctx context.Context, pagePkID int64, updateInput domain.PageUpdateInput) (*domain.Page, *domain.Error) {
+	var page = model.Page{}
+	if dbErr := r.store.DB().Where("pkid = ?", pagePkID).First(&page).Error; dbErr != nil {
+		return nil, domain.NewErr("Page not found", domain.BadRequestCode)
+	}
+	if updateInput.Name != nil {
+		page.Name = *updateInput.Name
+	}
+	if updateInput.ViewType != nil {
+		page.ViewType = updateInput.ViewType.String()
+	}
+	if updateInput.ParentPagePkID != nil {
+		page.ParentPagePkid = updateInput.ParentPagePkID
+	}
+	if updateInput.CoverImage != nil {
+		page.CoverImage = *updateInput.CoverImage
+	}
 
-// func (r *PageRepository) UpdatePageByID(ctx context.Context, pageID string, newPage domain.PageInput) (*domain.Page, *domain.Error) {
-// 	var page = model.Page{}
+	dbErr := r.store.DB().Clauses(clause.Returning{}).Select("*").Save(&page).Error
+	if dbErr != nil {
+		return nil, domain.ErrDatabaseMutation
+	}
 
-// 	dbErr := r.store.DB().Where("id = ?", pageID).First(&page).Error
-// 	if dbErr != nil {
-// 		return nil, domain.NewErr("Page not found", domain.BadRequestCode)
-// 	}
+	return docutils.TransformPageModelToDomain(
+		page,
+		nil,
+		nil,
+	), nil
+}
 
-// 	page.Name = newPage.Name
-// 	page.ViewType = newPage.ViewType
-// 	page.ParentPagePkid = newPage.ParentPagePkID
-// 	page.CoverImage = newPage.CoverImage
+func (r *DocRepository) CreatePage(ctx context.Context, pageInput domain.PageInput) (*domain.Page, *domain.Error) {
+	newPage := model.Page{
+		Name:           pageInput.Name,
+		CoverImage:     pageInput.CoverImage,
+		OrgPkid:        &pageInput.OrganizationPkID,
+		ParentPagePkid: pageInput.ParentPagePkID,
+		ViewType:       pageInput.ViewType.String(),
+	}
+	if pageInput.Document.JsonContent == "" {
+		pageInput.Document.JsonContent = "{}"
+	}
 
-// 	dbErr = r.store.DB().Clauses(clause.Returning{}).Select("*").Save(&page).Error
+	// Begin Tx
+	tx, doneTx := r.store.NewTransaction()
+	err := tx.DB().Create(&newPage).Error
+	if err != nil {
+		return nil, doneTx(err)
+	}
 
-// 	if dbErr != nil {
-// 		return nil, domain.ErrDatabaseMutation
-// 	}
+	document := model.Document{
+		JSONContent: &pageInput.Document.JsonContent,
+		PagePkid:    newPage.Pkid,
+	}
 
-// 	return pageutils.TransformPageModelToDomain(page, nil), nil
-// }
+	rerr := tx.DB().Create(&document).Error
+	if rerr != nil {
+		return nil, doneTx(err)
+	}
 
-// func (r *PageRepository) ArchivedPageByID(ctx context.Context, pageID string) (*domain.Page, *domain.Error) {
-// 	var page = model.Page{}
-// 	err := r.store.DB().Where("id = ?", pageID).First(&page).Error
-// 	if err != nil {
-// 		return nil, domain.NewErr("Page not found", domain.BadRequestCode)
-// 	}
-// 	now := time.Now()
-// 	page.ArchivedAt = &now
+	doneTx(nil)
+	// Commit Tx
 
-// 	err = r.store.DB().Clauses(clause.Returning{}).Select("*").Save(&page).Error
-// 	if err != nil {
-// 		return nil, domain.ErrDatabaseMutation
-// 	}
-// 	return pageutils.TransformPageModelToDomain(page, nil), nil
-// }
+	return docutils.TransformPageModelToDomain(
+		newPage,
+		[]domain.Page{},
+		docutils.TransformDocModalToDomain(document),
+	), nil
+}
 
-// func (r *PageRepository) GetPagesByNodeID(ctx context.Context, nodeIDs []string) ([]domain.Page, *domain.Error) {
-// 	var pages []model.Page
-// 	err := r.store.DB().Where("node_id IN ?", nodeIDs).Find(&pages).Error
-// 	if err != nil {
-// 		return nil, domain.ErrDatabaseQuery
-// 	}
+func (r *DocRepository) GetByID(ctx context.Context, pageID string) (*domain.Page, *domain.Error) {
+	var page model.Page
+	if dbErr := r.store.DB().Where("id = ?", pageID).First(&page).Error; dbErr != nil {
+		return nil, domain.ErrDatabaseQuery
+	}
 
-// 	var domainPages []domain.Page
-// 	for _, page := range pages {
-// 		domainPages = append(domainPages, *pageutils.TransformPageModelToDomain(page, nil))
-// 	}
+	var childPages []model.Page
+	if dbErr := r.store.DB().Where("parent_page_pkid = ?", page.Pkid).Find(&childPages).Error; dbErr != nil {
+		return nil, domain.ErrDatabaseQuery
+	}
 
-// 	return domainPages, nil
-// }
+	var doc model.Document
+	if dbErr := r.store.DB().Where("page_pkid = ?", page.Pkid).First(&doc).Error; dbErr != nil {
+		return nil, domain.ErrDatabaseQuery
+	}
 
-// func (r *PageRepository) BulkCreatePages(ctx context.Context, newPagesInput []domain.PageInput) ([]domain.Page, *domain.Error) {
-// 	var pages []model.Page
-// 	for _, page := range newPagesInput {
-// 		nodeID := &page.NodeID
-// 		if nodeID == nil {
-// 			uuid := uuid.NewString()
-// 			nodeID = &uuid
-// 		}
-// 		pages = append(pages, model.Page{
-// 			Name:           page.Name,
-// 			SpacePkid:      page.SpacePkID,
-// 			ViewType:       page.ViewType,
-// 			ParentPagePkid: page.ParentPagePkID,
-// 			NodeID:         nodeID,
-// 		})
-// 	}
-// 	err := r.store.DB().Create(&pages).Error
-// 	if err != nil {
-// 		return nil, domain.ErrDatabaseMutation
-// 	}
+	childPagesDomain := make([]domain.Page, len((childPages)))
+	for i := 0; i < len(childPages); i++ {
+		childPagesDomain[i] = *docutils.TransformPageModelToDomain(childPages[i], nil, nil)
+	}
 
-// 	var domainPages []domain.Page
-// 	for _, page := range pages {
-// 		domainPages = append(domainPages, *pageutils.TransformPageModelToDomain(page, nil))
-// 	}
+	return docutils.TransformPageModelToDomain(
+		page,
+		childPagesDomain,
+		docutils.TransformDocModalToDomain(doc),
+	), nil
+}
 
-// 	return domainPages, nil
-// }
+func (r *DocRepository) UpdateContent(ctx context.Context, pagePkID int64, jsonContent string) (*domain.Page, *domain.Error) {
+	var page = model.Page{}
+	if dbErr := r.store.DB().Where("pkid = ?", pagePkID).First(&page).Error; dbErr != nil {
+		return nil, domain.NewErr(dbErr.Error(), domain.BadRequestCode)
+	}
 
-// func (r *PageRepository) BulkArchivePages(ctx context.Context, pagePkIDs []int64) *domain.Error {
-// 	var pages []model.Page
-// 	for _, pagePkID := range pagePkIDs {
-// 		pages = append(pages, model.Page{
-// 			Pkid: pagePkID,
-// 		})
-// 	}
-// 	err := r.store.DB().Model(&pages).Update("archived_at", time.Now()).Error
-// 	if err != nil {
-// 		return domain.ErrDatabaseMutation
-// 	}
-// 	return nil
-// }
+	var doc model.Document
+	if dbErr := r.store.DB().Where("page_pkid = ?", pagePkID).First(&doc).Error; dbErr != nil {
+		return nil, domain.NewErr(dbErr.Error(), domain.BadRequestCode)
+	}
+	if jsonContent == "" {
+		jsonContent = "{}"
+	}
+	doc.JSONContent = &jsonContent
+	if dbErr := r.store.DB().Clauses(clause.Returning{}).Select("*").Save(&doc).Error; dbErr != nil {
+		return nil, domain.ErrDatabaseMutation
+	}
+
+	return docutils.TransformPageModelToDomain(
+		page,
+		nil,
+		docutils.TransformDocModalToDomain(doc),
+	), nil
+}
+
+func (r *DocRepository) Archive(ctx context.Context, pagePkID int64) (*domain.Page, *domain.Error) {
+	return nil, nil
+}
